idea finalized
