+++ lib/commons/color/get-background-color.js
@@ -1,6 +1,6 @@
 import incompleteData from './incomplete-data';
 import getBackgroundStack from './get-background-stack';
-// import getOwnBackgroundColor from './get-own-background-color';
+import getOwnBackgroundColor from './get-own-background-color';
 import elementHasImage from './element-has-image';
 import Color from './color';
 import flattenColors from './flatten-colors';
@@ -8,6 +8,7 @@ import flattenShadowColors from './flatten-shadow-colors';
 import getTextShadowColors from './get-text-shadow-colors';
 import getVisibleChildTextRects from '../dom/get-visible-child-text-rects';
 import { getNodeFromTree } from '../../core/utils';
+import { getStackingContext, stackingContextToColor } from './stacking-context';
 
 /**
  * Returns background color for element
@@ -47,45 +48,15 @@ export default function getBackgroundColor(
 }
 
 function _getBackgroundColor(elm, bgElms, shadowOutlineEmMax) {
-  let bgColors = getTextShadowColors(elm, { minRatio: shadowOutlineEmMax });
-  if (bgColors.length) {
-    bgColors = [{ color: bgColors.reduce(flattenShadowColors) }];
-  }
-
   const elmStack = getBackgroundStack(elm);
-  if (elmStack === null) {
+  if (!elmStack) {
     return null;
   }
 
   const textRects = getVisibleChildTextRects(elm);
-  const stackingContext = [];
-  const contextMap = new Map();
-
-  if (window.DEBUG) {
-    console.log(
-      JSON.stringify(
-        elmStack.map(node => node.cloneNode().outerHTML),
-        null,
-        2
-      )
-    );
-
-    console.log(
-      JSON.stringify(
-        elmStack.map(node => getNodeFromTree(node)._stackingOrder),
-        (key, value) => {
-          if (key === 'vNode') {
-            if (value?.props.nodeName === 'html') {
-              return 'html';
-            }
-            return value?.actualNode.cloneNode().outerHTML;
-          }
-
-          return value;
-        },
-        2
-      )
-    );
+  let bgColors = getTextShadowColors(elm, { minRatio: shadowOutlineEmMax });
+  if (bgColors.length) {
+    bgColors = [{ color: bgColors.reduce(flattenShadowColors) }];
   }
 
   // Search the stack until we have an alpha === 1 background
@@ -95,13 +66,11 @@ function _getBackgroundColor(elm, bgElms, shadowOutlineEmMax) {
 
     if (elementHasImage(bgElm, bgElmStyle)) {
       bgElms.push(bgElm);
-      if (window.DEBUG) {
-        console.log('elementHasImage');
-      }
+
       return null;
     }
 
-    // ignore fully transparent backgrounds
+    // Get the background color
     const bgColor = getOwnBackgroundColor(bgElmStyle);
     if (bgColor.alpha === 0) {
       continue;
@@ -112,9 +81,6 @@ function _getBackgroundColor(elm, bgElms, shadowOutlineEmMax) {
       bgElmStyle.getPropertyValue('display') !== 'inline' &&
       !fullyEncompasses(bgElm, textRects)
     ) {
-      if (window.DEBUG) {
-        console.log('elmPartiallyObscured');
-      }
       bgElms.push(bgElm);
       incompleteData.set('bgColor', 'elmPartiallyObscured');
 
@@ -122,60 +88,16 @@ function _getBackgroundColor(elm, bgElms, shadowOutlineEmMax) {
     }
 
     // store elements contributing to the bg color.
-    let alpha = 1;
-    const bgVNode = getNodeFromTree(bgElm);
-    const stackingOrder = bgVNode._stackingOrder.filter(({ vNode }) => !!vNode);
-    stackingOrder.forEach(({ vNode }, index) => {
-      const ancestorVNode = stackingOrder[index - 1]?.vNode;
-      const context = addToStackingContext(contextMap, vNode, ancestorVNode);
-      alpha *= context.opacity;
-
-      if (index === 0 && !contextMap.get(vNode)) {
-        stackingContext.unshift(context);
-      }
-      contextMap.set(vNode, context);
-    });
-
-    const ancestorVNode = stackingOrder[stackingOrder.length - 1]?.vNode;
-    const context = addToStackingContext(contextMap, bgVNode, ancestorVNode);
-    if (!stackingOrder.length) {
-      stackingContext.unshift(context);
-    }
-    context.bgColor = bgColor;
-    alpha *= context.opacity * bgColor.alpha;
-
-    if (bgColor.alpha !== 0) {
-      bgElms.push(bgElm);
-    }
+    bgElms.push(bgElm);
 
     // Exit if the background is opaque
-    if (alpha === 1) {
+    if (bgColor.alpha === 1) {
       break;
     }
   }
-  if (window.DEBUG) {
-    console.log(
-      JSON.stringify(
-        stackingContext,
-        (key, value) => {
-          if (key === 'vNode') {
-            if (value?.props.nodeName === 'html') {
-              return 'html';
-            }
-            return value?.actualNode.cloneNode().outerHTML;
-          }
 
-          return value;
-        },
-        2
-      )
-    );
-  }
-  bgColors = stackingContext.map(flattenStack).concat(bgColors);
-  if (window.DEBUG) {
-    console.log('bgColors:', JSON.stringify(bgColors, null, 2));
-    console.log('includes body:', elmStack.includes(document.body));
-  }
+  const stackingContext = getStackingContext(elm, elmStack);
+  bgColors = stackingContext.map(stackingContextToColor).concat(bgColors);
 
   const pageBgs = getPageBackgroundColors(
     elm,
@@ -191,10 +113,6 @@ function _getBackgroundColor(elm, bgElms, shadowOutlineEmMax) {
   // order (background to foreground order) to produce the correct
   // result.
   // @see https://github.com/dequelabs/axe-core/issues/2924
-
-  if (window.DEBUG) {
-    console.log('bgColors (2):', JSON.stringify(bgColors, null, 2));
-  }
   const blendedColor = bgColors.reduce((bgColor, fgColor) => {
     return flattenColors(
       fgColor.color,
@@ -209,79 +127,6 @@ function _getBackgroundColor(elm, bgElms, shadowOutlineEmMax) {
     blendedColor.color instanceof Color ? blendedColor.color : blendedColor,
     new Color(255, 255, 255, 1)
   );
-}
-
-function createStackingContext(vNode) {
-  return {
-    vNode: vNode,
-    opacity: parseFloat(vNode?.getComputedStylePropertyValue('opacity') ?? 1),
-    bgColor: new Color(0, 0, 0, 0),
-    blendMode: normalizeBlendMode(
-      vNode?.getComputedStylePropertyValue('mix-blend-mode')
-    ),
-    descendants: []
-  };
-}
-
-function addToStackingContext(contextMap, vNode, ancestor) {
-  const context = contextMap.get(vNode) ?? createStackingContext(vNode);
-  const ancestorContext = contextMap.get(ancestor);
-  if (
-    ancestorContext &&
-    ancestor !== vNode &&
-    !ancestorContext.descendants.includes(context)
-  ) {
-    ancestorContext.descendants.unshift(context);
-  }
-
-  return context;
-}
-
-function reduceToColor(bgContext, fgContext) {
-  let bgColor;
-  if (bgContext instanceof Color) {
-    bgColor = bgContext;
-  } else {
-    bgColor = flattenStack(bgContext).color;
-  }
-
-  const fgColor = flattenStack(fgContext).color;
-  return flattenColors(fgColor, bgColor, fgContext.blendMode);
-}
-
-function flattenStack(context) {
-  if (!context.descendants?.length) {
-    const color = context.bgColor;
-    color.alpha *= context.opacity;
-    return {
-      color,
-      blendMode: context.blendMode
-    };
-  }
-
-  const fgColor = context.descendants.reduce(
-    reduceToColor,
-    createStackingContext()
-  );
-  const color = flattenColors(fgColor, context.bgColor);
-  color.alpha *= context.opacity;
-  return {
-    color,
-    blendMode: context.blendMode
-  };
-}
-
-function getOwnBackgroundColor(vNode) {
-  const bgColor = new Color();
-  if (vNode.getPropertyValue) {
-    bgColor.parseString(vNode.getPropertyValue('background-color'));
-  } else {
-    bgColor.parseString(
-      vNode.getComputedStylePropertyValue('background-color')
-    );
-  }
-
-  return bgColor;
 }
 
 /**
@@ -320,6 +165,7 @@ function fullyEncompasses(node, rects) {
 function normalizeBlendMode(blendmode) {
   return !!blendmode ? blendmode : undefined;
 }
+
 /**
  * Get the page background color.
  * @private
